requirements.txt
pygame
pyserial
pillow
sv-ttk
